♻️ A9YYYQ code: deduplicate hook task context
