+++ .agentplane/tasks/202605091753-A9YYYQ/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202605091753-A9YYYQ/hook-context",
+  "created_at": "2026-05-09T18:37:16.398Z",
+  "head_sha": "31abc088e93810626bbf2ac1abd979c42abb6194",
+  "last_verified_at": "2026-05-09T18:37:16.060Z",
+  "last_verified_sha": "31abc088e93810626bbf2ac1abd979c42abb6194",
+  "schema_version": 1,
+  "task_id": "202605091753-A9YYYQ",
+  "updated_at": "2026-05-09T18:37:16.398Z",
+  "verify": {
+    "status": "pass"
+  }
+}
